Added an OSAL layer
The application has a lot of code that is OS specific.
Currently the application is targetting the Linux
operating system, but isolate the OS specific functions
such that in future would be easier to port the application
on other operating systems. The separation is not 100%, the
source file containg the main function is also OS dependent, but
it was kept standalone, independent of the OS layer for clarity.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>