Add support for closing the device
When the device provisioning finished the user may
opt for closing the device. The operation is
ireversible and after this step the device will allow
booting signed images only. The step can be skipped from
the configuration file, the user can use a command line
argument to later close the device.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>