Delete the assets file after provisioning
After provisioning, the user may choose to delete the assets file by
configuring this behavior in the configuration file.

This option is particularly useful when boot-time provisioning is
enabled, but the user does not want the provisioning process to repeat
on every boot. By removing the assets file after the initial
provisioning, the system ensures that provisioning only occurs once.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>