Use psa_its_remove for EdgeLock 2GO data
EdgeLock 2GO blobs are imported using psa_import_key,
regardless of whether they contain cryptographic keys
or general data. However, when removing these
secure objects from storage, the appropriate deletion
function depends on the object type:
* Use psa_destroy_key for keys.
* Use psa_its_remove for data.

Since the object type associated with a specific ID is not known in
advance, the following strategy is implemented:
* Attempt to delete the object using psa_destroy_key.
* If psa_destroy_key returns PSA_ERROR_INVALID_HANDLE, indicating the ID
does not correspond to a key, then call psa_its_remove.

This approach ensures that the correct deletion function is eventually
called, regardless of whether the object is a key or data.
Signed-off-by: Diana Craciun <diana.craciun@nxp.com>