Allow the usage of multiple assets file
The product based provisioning may use two assets file:
one in open state and one in close state. Extend the
interface to support such a scenario. Currently the product
based provisioning is allowed only in open state for S400,
so the application will limit the number of files to 1.
However, the list is defined to allow extensions in the
future without the need to change the interface.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>