Correct the offset calculation formula
The offset calculation formula did not take into account
that, at the time the offset of a blob in the stream is
calculated, the stream cursor is at the end of the magic
TLV field in the next blob, so the calculated offsets
differed by one magic TLV field length from the correct
value. The offset of the last blob was calculated correctly.
Changed the formula to calculate the correct offset for
all blobs.

Signed-off-by: Florin Pavelescu <florin.pavelescu@nxp.com>