Replace indirect with offline
Indirect flow was renamed to Provisioning by Proxy.
However from the point of view of the application
there is not much of a difference between provisioning
by proxy and product based provisioning. The same
section of the configuration file will be used for
both flows. The code was also changed in order to
avoid confusions.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>