Remove the Yocto recipe from the application
The Yocto recipe is removed from the application folder
as the application is now part of the BSP.
Also move the Yocto layer under documentation as is
not actually part of the documentation, but an example
offered to run the application at boot. I could have
added all the details in the documentation, but it
is rather difficult to maintain the documentation steps
than an actual layer that can be easily used to test.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>