Report an error if OEM key is found for provisioning via proxy
If the user has set in the configuration file provisioning
type to individual, the application will report an error
if the OEM key is found.
If no configuration was set in the configuration file, the
application will try to import all the blobs found starting
with the OEM key (if found) and SRKH (if found).

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>